// AngularMain/src/app/modules/models/CoaDto.ts
    export interface CoaDto {
        tenentId: number;
        locationId: number;
        accountId: number;
        familyId: number;
        headId: number;
        subHeadId: number;
        accountNumber: number;
        accountName: string;
        arabicAccountName: string;
        otherAccountName?: any;
        accountTypeId: number;
        isActive: boolean;
        createdBy?: any;
        createdDate?: any;
        updatedBy?: any;
        updatedDate?: any;
        crupid?: any;
        bankAccountNo?: any;
        subSubHeadId: number;
    }